Factor out parsring of observation response
